+++ docs/changelog.rst
@@ -15,6 +15,8 @@ Adds **Foundation for site 6** support, version 5 support is still available for
 * Fixed switches for Foundation-6;
 * Fixed button group for Foundation-6;
 * Fixed ``InlineJustifiedField`` for Foundation-6;
+* Fixed error messages for Foundation-6;
+* Added ``base.Callout`` element;
 * ``.help-text`` is allways a ``<p>`` in Foundation6 (does not have real meaning in Foundation5);
 * Added ``crispy_forms_foundation.templatetags.crispy_forms_foundation_field`` to re-implement ``crispy_field`` filter so we can have the right input field error for Foundation-6;